requirements.txt
 
fastapi
uvicorn
openai
websockets
python-dotenv